pyproject.toml
[tool.poetry]
name = "microtechnology-skeleton"
version = "0.4.0"
description = "Template for Generative Microtechnologies"
authors = ["Generative Engineering <degenerates@generative.vision>"]
packages = [
    { include = "cube", from = "src" },
]

[[tool.poetry.source]]
name = "gitlab"
url = "https://gitlab.com/api/v4/groups/14110432/-/packages/pypi/simple"
secondary = true

[tool.poetry.dependencies]
python = "^3.9"
generative-fabric-definitions = "*"
cadquery = "^2.3.1"
cadquery-ocp = "=7.7.0"

[tool.poetry.dev-dependencies]
mypy = "^1.4.1"
pytest = "^7.1.2"
black = "^23.7.0"
ruff = "^0.0.280"

[tool.poetry.scripts]
server = 'generative.fabric.http.__main__:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Linter config

[tool.black]
line-length = 99

[tool.ruff]
line-length = 99
select = ["E", "F", "B", "Q", "C", "N"]
target-version = "py39"

[tool.mypy]
strict = true
files = "src,tests"
warn_unused_configs = true
namespace_packages = true
warn_return_any = true
show_error_context = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
# Too noisy to have `-> None` on every test
# and to type some Generator fixtures is hellish
# Unfortunately you need to turn *both* these off for this to work.
allow_incomplete_defs = true
allow_untyped_defs = true
